Fix the content boxes from being clumped together to being spread out
The problem I noticed when using the developer tools using my browser was that the div for the contents had a height that was fit to the child div's, instead of the whole viewport. Hence why justify property wasn't visibly doing anything.

First I needed to add a custom height for the div. But it still wasn't doing anything.

Then I solution was using `100vh` instead of `100%` for the height of the content div.